requirements.txt
pydantic==1.8.2
fastapi==0.67.0
pymongo==3.12.0
psutil==5.8.0
requests==2.26.0
python-dateutil==2.8.2
pyjwt==2.3.0
dnspython==2.1.0